Move extract_entity_id to service helpers
