+++ homeassistant/helpers/__init__.py
@@ -3,7 +3,6 @@ Helper methods for components within Home Assistant.
 """
 import re
 
-from homeassistant.loader import get_component
 from homeassistant.const import (
     ATTR_ENTITY_ID, CONF_PLATFORM, DEVICE_DEFAULT_NAME)
 from homeassistant.util import ensure_unique_string, slugify
@@ -20,25 +19,6 @@ def generate_entity_id(entity_id_format, name, current_ids=None, hass=None):
 
     return ensure_unique_string(
         entity_id_format.format(slugify(name.lower())), current_ids)
-
-
-def extract_entity_ids(hass, service):
-    """
-    Helper method to extract a list of entity ids from a service call.
-    Will convert group entity ids to the entity ids it represents.
-    """
-    if not (service.data and ATTR_ENTITY_ID in service.data):
-        return []
-
-    group = get_component('group')
-
-    # Entity ID attr can be a list or a string
-    service_ent_id = service.data[ATTR_ENTITY_ID]
-
-    if isinstance(service_ent_id, str):
-        return group.expand_entity_ids(hass, [service_ent_id])
-
-    return [ent_id for ent_id in group.expand_entity_ids(hass, service_ent_id)]
 
 
 def validate_config(config, items, logger):
